+++ src/main/java/dooya/see/application/post/CommentModifyService.java
@@ -4,6 +4,9 @@ import dooya.see.application.post.provided.CommentManager;
 import dooya.see.application.post.required.CommentRepository;
 import dooya.see.application.post.required.PostRepository;
 import dooya.see.domain.post.*;
+import dooya.see.domain.post.exception.CommentNotFoundException;
+import dooya.see.domain.post.exception.PostNotFoundException;
+import dooya.see.domain.post.exception.UnauthorizedCommentAccessException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
