+++ src/main/java/dooya/see/domain/post/Post.java
@@ -1,6 +1,7 @@
 package dooya.see.domain.post;
 
 import dooya.see.domain.post.event.*;
+import dooya.see.domain.post.exception.InvalidPostStatusTransitionException;
 import dooya.see.domain.shared.AbstractAggregateRoot;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
