+++ src/main/java/dooya/see/domain/post/exception/PostNotFoundException.java
@@ -1,4 +1,4 @@
-package dooya.see.domain.post;
+package dooya.see.domain.post.exception;
 
 public class PostNotFoundException extends RuntimeException {
     public PostNotFoundException(String message) {
