+++ src/main/java/dooya/see/domain/post/exception/UnauthorizedPostAccessException.java
@@ -1,4 +1,4 @@
-package dooya.see.domain.post;
+package dooya.see.domain.post.exception;
 
 public class UnauthorizedPostAccessException extends RuntimeException {
     public UnauthorizedPostAccessException(String message) {
