+++ src/test/java/dooya/see/application/post/provided/CommentManagerTest.java
@@ -4,6 +4,10 @@ import dooya.see.SeeTestConfiguration;
 import dooya.see.application.post.required.CommentRepository;
 import dooya.see.application.post.required.PostRepository;
 import dooya.see.domain.post.*;
+import dooya.see.domain.post.exception.CommentNotFoundException;
+import dooya.see.domain.post.exception.InvalidCommentStatusException;
+import dooya.see.domain.post.exception.PostNotFoundException;
+import dooya.see.domain.post.exception.UnauthorizedCommentAccessException;
 import jakarta.persistence.EntityManager;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
