+++ src/test/java/dooya/see/application/post/provided/PostManagerTest.java
@@ -3,6 +3,9 @@ package dooya.see.application.post.provided;
 import dooya.see.SeeTestConfiguration;
 import dooya.see.application.post.required.PostLikeRepository;
 import dooya.see.domain.post.*;
+import dooya.see.domain.post.exception.InvalidPostStatusTransitionException;
+import dooya.see.domain.post.exception.PostNotFoundException;
+import dooya.see.domain.post.exception.UnauthorizedPostAccessException;
 import jakarta.persistence.EntityManager;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
