+++ src/test/java/dooya/see/domain/post/CommentTest.java
@@ -1,6 +1,8 @@
 package dooya.see.domain.post;
 
 import dooya.see.domain.post.event.*;
+import dooya.see.domain.post.exception.EmptyCommentUpdateException;
+import dooya.see.domain.post.exception.InvalidCommentStatusException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
